+++ ggml/src/ggml-gemmini/ggml-gemmini-util.h
@@ -16,6 +16,25 @@
 #include <set>
 #include <cstring>
 
+struct ggml_backend_gemmini_context;
+#include "bench_tensor/gemmini_bench_tensor.h"
+
+struct ggml_backend_gemmini_context
+{
+    int n_threads = GGML_DEFAULT_N_THREADS;
+    std::unique_ptr<char[]> work_data;
+    size_t work_size = 0;
+    std::map<ggml_tensor *, ggml_tensor *> bias_map;
+    std::map<const ggml_tensor *, std::unique_ptr<zerogod::BenchTensor<int8_t>>> tensor_cache;
+    struct ggml_context *tmp_ctx = nullptr;
+    void *arena = nullptr;
+    bool tmp_ctx_initialized = false;
+
+#ifndef GGML_USE_OPENMP
+    std::vector<std::future<void>> tasks;
+#endif
+};
+
 #ifndef PRINT_TILE
 #define PRINT_TILE 0
 #endif
